requirements.txt
flask
pymongo
dnspython
bcrypt
pyjwt
flask-cors
python-dotenv